+++ src/web/assets/input/src/scripts/Input.js
@@ -80,6 +80,7 @@ export default Garnish.Base.extend({
   _siteId: null,
   _visibleLayoutElements: {},
   _newBlockCount: 0,
+  _creatingBlock: false,
 
   init (settings = {}) {
     settings = Object.assign({}, _defaults, settings)
@@ -1079,14 +1080,29 @@ export default Garnish.Base.extend({
     // Remove [blocks] from namespace
     NS.enter(Array.from(this.getNamespace()).slice(0, -1))
     if (elementEditor) {
-      // First ensure we're working with drafts for all elements leading up
-      // to this field’s element
+      // Ensure we're working with an owner draft if drafts haven't been disabled
       await elementEditor.setFormValue(
         NS.toFieldName(),
         '*'
       )
     }
     NS.leave()
+  },
+
+  _setCreatingBlock () {
+    if (this._creatingBlock) {
+      return false
+    }
+
+    this._creatingBlock = true
+    this.$container.find('.ni_buttons').addClass('disabled')
+
+    return true
+  },
+
+  _unsetCreatingBlock () {
+    this._creatingBlock = false
+    this.$container.find('.ni_buttons').removeClass('disabled')
   },
 
   _addSpinnerAfter (block) {
@@ -1136,6 +1152,11 @@ export default Garnish.Base.extend({
   },
 
   async _duplicate (data, block) {
+    if (!this._setCreatingBlock()) {
+      console.warning('Tried to create a new block during the creation of another new block.')
+      return
+    }
+
     data.unsavedIds = this.getUnsavedIds()
     try {
       this.$form.data('elementEditor')?.pause()
@@ -1188,10 +1209,16 @@ export default Garnish.Base.extend({
     } finally {
       this._removeSpinner()
       this.$form.data('elementEditor')?.resume()
+      this._unsetCreatingBlock()
     }
   },
 
   async '@newBlock' (e) {
+    if (!this._setCreatingBlock()) {
+      console.warning('Tried to create a new block during the creation of another new block.')
+      return
+    }
+
     await this._setDirty()
     const elementEditor = this.getElementEditor()
 
@@ -1256,6 +1283,7 @@ export default Garnish.Base.extend({
       Craft.cp.displayError(error.message)
     } finally {
       elementEditor?.resume()
+      this._unsetCreatingBlock()
     }
   },